+++ grails-app/services/io/xh/hoist/monitor/MonitorReportService.groovy
@@ -84,7 +84,7 @@ class MonitorReportService extends BaseService {
     }
 
     private MonitorConfig getConfig() {
-        new MonitorConfig(configService.getMap('xhMonitorConfig'))
+        configService.getTypedConfig(MonitorConfig)
     }
 
     Map getAdminStats() {[